[Compiler] Extract script naming from NameUtils into core/names
The script/snippet target-class-name derivation moved from 'NameUtils'
into a new focused 'ScriptNames' object in core/names, with 'NameUtils'
delegating to it.


^KT-89154
